finish with clean architecture

+++ app/src/main/java/com/sobarna/smartcalculatorapp/data/entity/CalculatorEntity.kt
@@ -9,7 +9,7 @@ data class CalculatorEntity(
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo("id")
-    val id: Int = 0,
+    val id: Long = 0,
 
     @ColumnInfo("input")
     val input:String?,
